requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
auth0-python==3.14.0
bcrypt==3.1.7
boto3==1.16.47
botocore==1.19.47
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-db-logger==0.1.10
django-debug-toolbar==3.2
django-extensions==3.1.0
django-jsonfield==1.4.1
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-ses==1.0.3
django-smtp-ssl==1.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.4
future==0.18.2
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-core==1.6.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
idna==2.10
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
mccabe==0.6.1
numpy==1.20.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
Pillow==7.0.0
pipenv==2020.11.15
protobuf==3.15.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rest-framework-auth0==0.6.4
rsa==4.7.2
s3transfer==0.3.3
six==1.14.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
xlrd==2.0.1
